Add the GitLab CI as well
Well, that's where the template was first made, so might as well add that in so I have both CIs done properly.

The only caveat is - the key (and keystore) used by GitHub Actions is *different* than the one used by GitLab CICD
but that's of less concern.

Just adding it here so that updating the template back is easier.