CI/CD setup (#1)
* Add CI/CD action to build APK

Using the following github secrets:
- KEYSTORE (base64 of key store)
- STOREPASSWORD (password for the keystore)
- KEYALIAS (key alias used for signing)
- KEYPASSWORD (password for the key alias)

Alright, we can start working on the first screen's request/response adapters and the UI now :)

EDIT: whoops, almost forgot test and test coverage reports.

* Try to fix CI/CD

* And try to fix it again ...

I dont know why "app" is missing but... in case it is - lets create it first.
But it shouldn't be missing, something's very wrong.

* Aaaa.... fix CI/CD by adding the forgotten checkout step

Ok, so i forgot the checkout step :D

well, that explains why 'app' was missing. so just avoid "mkdir app" now since it's surely
going to be there.

* Build Debug before Release, consolidate secrets used for the signing key

Ok, it should hopefully work out now, if the problem isn't related to gradlew
and the step used shouldn't make it a problem?

* Push gradle wrapper

The step is not setting it up properly, and the whole deal with gradle having to be
there but normally *shouldnt* be there (because God forbid, users can have different gradle)
and can't be fetched nicely without actual additional information on which gradle to get ...

It's just not worth it, lets juts do an anti-pattern here and push the gradle wrapper so CI/CD
can do it's job fine.

* Make gradlew executable in CI/CD workflow

* Fix build.gradle problem with File

* Change moving output/apk/**/*debug*.apk to what we find instead

-          mv app/build/outputs/apk/**/debug/*debug*.apk ./upload
-          mv app/build/outputs/apk/**/release/*release*.apk ./upload
+          find app/build/outputs/apk -name '*debug*.apk' -exec mv {} ./upload \;
+          find app/build/outputs/apk -name '*release*.apk' -exec mv {} ./upload \;

This says it all. Instead of looking for wildcard files in wildcard folders, lets just `find` them instead
and copy that over.

However i'm mainly suspecting that the build isn't succeeding properly, better safe than sorry...
the same gradlew command works fine on my machine :)

EDIT:
also, add validation after assembleDebug and assembleRelease to look for built APK
and print it's location. just to be sure we're building something.

* Add "cache gradle" and "run tests" CI steps

OK, the "cache gradle" is mainly because I see that these "build" steps are not
using cache, even though org.gradle.caching is set in gradle.properties.

BUT - i've also added a "run tests" step, because the latest CI/CD fix has failed
due to not finding test reports to move to the artefact (because tests weren't ran
and they are hooked into to generate reports as well ...... :/ )

* Few more CI/CD changes

1) let's also add "develop" branch to trigger CI/CD
2) lets restore the cached gradle in the "cache gradle" step
3) change the "run tests" to "build and run tests" to execute ./gradlew build test instead of just ./gradlew test

* Make action run on every push to any branch as well as when tags are created

* Consolidate spacing

* Add the GitLab CI as well

Well, that's where the template was first made, so might as well add that in so I have both CIs done properly.

The only caveat is - the key (and keystore) used by GitHub Actions is *different* than the one used by GitLab CICD
but that's of less concern.

Just adding it here so that updating the template back is easier.