GetRepositoriesResponseAdapter and it's underlying GSON json model data classes
I've decided that the search api to populate the first page with results will be called
"get repositories". The API call is yet to be implemented, however here's the necessary
ResponseAdapter for it. Next up - request adapter and the NetworkingUseCase to make it work.

Then the UI to tie it all together.