Added two new composables
First one is GenericLazyColumnWithOverscroll which is, as it's name suggest,
a LazyList that takes a list of items of generic type Type, a onOverscrollCallback and the Composable
to use for the items themselves.

The second one is LoadImageFromUrl composable which uses Coil to load an image from a given URL.

Lets see if they actually work once we put these two pieces into their proper place in the FirstScreen.