Whoops, fix the overscroll calculation ...
A short debug sesh revealed that I was wrong, and that I should actually check the other way :D

an additional improvement was put in to only invoke the callback once instead of N times as well.

That being said, the repo has to be done *NOW* because the items are being replaced with the next
"page"'s items as returned by the API.

Perhaps the same GitRepositoriesRepository can function as both the Favorites and Repositories repo.
In fact, the more I think about it, the more it should.

so git on it.