Room DB is in da haus
Added the necessary Database (extends RoomDatabase) class, with it's
entity list containing just GitRepository for the time being.

It also has the DatabaseConverters, which so far only cover converting
the Instant to/from String

DatabaseNames (to be renamed) contain the constants, such as version,
database file name and table name(s)

And finally, the GitRepositoryDao is the Dao class covering the insertion/
fetching of GitRepository to/from the database table itself.

EDIT: renamed DatabaseNames to DatabaseConstants and made them also hold the DB version