More database updates - favorites table
At first, I thought about just adding a 'isFavorite' to GitRepository, but then it dawned on me
that whenever the app is restarted, and new data is pulled from the web, that whatever we favorited
earlier would be overwritten by new results (from the web) and that we would get nothing out of all
that gymnastics.

So, a new "favorites" table is in order, that will share the same key as the GitRepository's primary key.

That way, we'll have a one-to-many which is really one-to-zero_or_one relationship between the
"git_repositories" and "favorites" tables.

Also, a new Favorite domain class, the FavoriteDao, and providing the FavoriteDao thru the AppModule.
A few more DatabaseConstants were added to streamline referncing the primary key ID.