Fix a very ugly database-related bug ...
Namely, it's as if the "favorites" table wasn't being persisted. Or working at all ...

But after some shamefully long digging around, it actually became clear that since we're
pulling new API results on every app restart, and GitRepoDao is REPLACING them - each REPLACE
is really a DELETE followed by an INSERT.

This, coupled with the Favorite's ForeignKey declaration, to perform ForeignKey.CASCADE in it's
onDelete action (invoked when the parent is deleted) is causing a full "favorites" table wipe on every
app restart.

So... This replacing needs to become a bit smarter. And that will most likely happen in the Repo layer.