Added FuzzyFilterer and a test for it
Whether this will be used by a UseCase or the GitReposRepository itself is yet to be decided ...

However, if the repo emits new values, they won't just go and re-execute the usecase.
If the repo uses them, that might work I suppose, but whether it's CLEAN clean is something to wonder about tomorrow.

Maybe the presenter can use them? But that sort of violates the idea that presenters shouldn't have any business logic
in them ... hmm hmm hmm...

In any case, since the FuzzyFilterer still mentions `File`s everywhere, that needs to be adapted to either GitRepository or just String