Fix database problem with item/repo ordering (preserving ordering from API as they came in)
The problem with the GitRepository's old primary key (id) is that it's outside
of our control, so we had to put in a new autogenerated primary key (_databaseId)
and make the 'old' primary key (id) be unique so we can still use the foreign key
relation with it.

The favorites still seem to function fine, so this is as good of a fix as any.
And it seems to be the right one too to me, and my humble DB knowledge from uni.