Extend the hacky TransitionalDataRepository with 'clicked url'
It didn't even have to be this way, the presenter itself could have hold onto it...
but anyways. the presenter got the GitRepositoriesRepository as well, so favoring works
and is visible on both screens.

So, with favoring done, and opening the URL done, the only thing left to do is -
parsing the "contributors_url" response, and drawing them on the screen besides
making a repository for them.