Add LiteralUrlApiCall capability to the networking framework
Alright, so these LiteralUrlApiCalls will target a literal URL provided by the caller/user.
We will piggyback off the querymap once again because, why not. It's useful only for QueriedApiCalls,
and since they have to be there because of *lack of foresight reasons*, might as well use it for this.

So now we can probably go ahead and finish the GetCollaboratorsUseCase and see how that one works out
in practice on the SecondScreen and it's presenter.

Speaking of which, did i actually finish the GetRepositoryDetails thing? The data I see coming from it
is barely different from the one we get from the first API call, but lets also implement it for completeness sake.