Collaborators are showing - but only after pressing the button.
Ok, so it's looking better and better BUT

1) need to make a ContributorsRepository, and the whole ContributorsDao
it doesn't need any ForeignKeys like the repository one, so just a simple Dao in the database
would do to save these collaborators
2) the GetCollaboratorsFromRepoDetailsUseCase needs a parameter to decide whether it's gonna grab
collaborators or contributors - I wanna show both
3) the button on the second screen needs to go away, and the whole usecase execution needs to happen
automagically. and populating both rows of icons. they can both write to the same database, why not.
4) The GitRepoView needs love for the *other* thing that didn't get changed to Column
5) the SecondScreen needs to pass the favorite view for the *other* thing