Second screen (#4)
* Initial implementation of getRepoDetails API call (and usecase) which likely won't work

* Temporarily hook the thing that doesn't work in hte second screen's presenter

* Try a dirty hack with TwoPathApiCall ...

* Fix GitRepository domain class

Ok, so we've been parsing the repository fullName as 'name', which is bad.
Differentiate between teh fullName and name, and just use 'name' for repoName from now on,
and leave a possibility to use the fullName as well (which is what the repo filtering should use)

* The TwoPathApiCall idea seems to have worked out fine

Alright, so now it needs to be tightened up better, and made more generic ...
Try to have a sealed class PathApiCall<...> that extends ApiCall, and make the
inner OneParamPathCall, TwoParamPathCall and the like somehow fit into that.

But dont waste too much time on it, the second screen needs to be finalized and there's
the whole third screen to finish, and I'm not yet getting where these collaborators should
be favored...

* Tighten up type-safety around the PathApiCall and it's subclasses

* Refactor something on FirstScreen - dunno how this slipped

* Fix a crash with parsing instants on second screen by initializing AndroidThreeTen in application class

* Extend the hacky TransitionalDataRepository with 'clicked url'

It didn't even have to be this way, the presenter itself could have hold onto it...
but anyways. the presenter got the GitRepositoriesRepository as well, so favoring works
and is visible on both screens.

So, with favoring done, and opening the URL done, the only thing left to do is -
parsing the "contributors_url" response, and drawing them on the screen besides
making a repository for them.

* Make the SecondScreen actually go to URL and show the repo since i didn't commit this until now ...

* GetCollaborators response adapter, domain class and the wrapper ResponseDomainData class

Well, it's obvious that we'll need to shoot an API call to get the collaborators_url and parse all this
into domain GitCollaborators, and for that, I'll actually need a new type of ApiCall - a DirectUrlApiCall

So before I get on that, lets save what's made so far.

* Add LiteralUrlApiCall capability to the networking framework

Alright, so these LiteralUrlApiCalls will target a literal URL provided by the caller/user.
We will piggyback off the querymap once again because, why not. It's useful only for QueriedApiCalls,
and since they have to be there because of *lack of foresight reasons*, might as well use it for this.

So now we can probably go ahead and finish the GetCollaboratorsUseCase and see how that one works out
in practice on the SecondScreen and it's presenter.

Speaking of which, did i actually finish the GetRepositoryDetails thing? The data I see coming from it
is barely different from the one we get from the first API call, but lets also implement it for completeness sake.

* Left a TODO to actually finish the GetRepositoryDetails response parsing and it's domain class

* fucking hell fucking copypasting

Ok, so since I based the LiteralUrlApiCall off of QueriedApiCall and was consistently
hitting into a java.lang.IllegalArgumentException: Failed requirement

Tried stepping with a debugger, wasn't even hitting a breakpoint, but I figured "hey,
maybe we can't break inside a future being executed somewhere" blahblahblah ...

fucking hell, I used the wrong enum for the ApiCallType, of course it wasn't hitting
the breakpoint where the code went somewhere else...

on the bright side, the GetCollaboratorsUseCase worked right off the bat and is returning
items just fine.

* Add everything we need to have the GetCollaboratorsUseCase working fine

* Fix the GetRepositoryDetailsResponseAdapter, replace teh two usecases with just one in SecondScreenViewModel

Besides that, the RepositoryDetails class is very lean and so far we're not putting *anything* significant in it,
and I'm having it simply because teh task requires me to call this API. We have all teh information we need(ed) from
the first API call, but - lets just grab the collaboratorUrl and contributorUrl from it and show them on the screen.

EDIT: remove the unnecessary imports

* Fix GetRepositoryDetailsUseCase - it was emitting an empty object because i forgot to fix that ...

* Fix alot of stuff everywhere and expand on GitRepoView

* Hmm, the lazyrow with the contributors isn't showing ...

* Collaborators are showing - but only after pressing the button.

Ok, so it's looking better and better BUT

1) need to make a ContributorsRepository, and the whole ContributorsDao
it doesn't need any ForeignKeys like the repository one, so just a simple Dao in the database
would do to save these collaborators
2) the GetCollaboratorsFromRepoDetailsUseCase needs a parameter to decide whether it's gonna grab
collaborators or contributors - I wanna show both
3) the button on the second screen needs to go away, and the whole usecase execution needs to happen
automagically. and populating both rows of icons. they can both write to the same database, why not.
4) The GitRepoView needs love for the *other* thing that didn't get changed to Column
5) the SecondScreen needs to pass the favorite view for the *other* thing

* Add capability for grabbing collaborators too - but since many repos don't have it, i'll get rid of it

* Get rid of showing/grabbing collaborators - since many projects don't have parsable collaborators

* UI fixes on CollaboratorView

* Mark GitCollaborator as entity and make the CollaboratorsDao with GitCollaboratorsRepository interface

* GitCollaboratorDao additions, GitCollaboratorsRepositoryImpl, binding in AppModule

* Fix CollaboratorsDao and make GitCollaborator unique by it's "login" column

Also, change the COLLABORATOR_PRIMARY_KEY constant to "login" since that's the key we'll be using mostly
for queries ...

* Fix collaborator DAO and repository, tie in with the presenter and the screen

Ok, now we can enter repo details and see the collaborators we already favored.
Next up:
1) make the fetching of repo details automagical, and remove teh "get repo details" button
2) start working on the third screen since i think the secondscreen is done.

* Make repo details fetching automagical

* Remove THIRD_SCREEN from SecondScreen's navigation events, that screen might make most sense to enter from the first screen

* Address PATH api/networking CR comments

* Address the rest of CR comments